apps: move block volume operations tests to new file
